% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{as.percent}
\alias{as.percent}
\title{Print a numeric as a nice percent}
\usage{
as.percent(x, dig = 1)
}
\arguments{
\item{x}{a numeric between 0 and 1}

\item{dig}{number of digits to print}
}
\value{

}
\description{
Print a numeric as a nice percent
}
