% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ci}
\alias{ci}
\title{Print a confidence interval}
\usage{
ci(x, ci = 0.95, dig = 2)
}
\arguments{
\item{x}{sample}

\item{ci}{width of the interval}

\item{dig}{number of digits to print}
}
\value{

}
\description{
Print a confidence interval
}
