% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is.monotonic}
\alias{is.monotonic}
\title{Check if a vector is monotonic}
\usage{
is.monotonic(x, type = "growing")
}
\arguments{
\item{x}{numerical vector}

\item{type}{matched to `"growing"` or `"decreasing"`}
}
\value{
TRUE/FALSE
}
\description{
Check if a vector is monotonic
}
