% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{or}
\alias{or}
\title{Calculate log odds ratio}
\usage{
or(a, b, c, d)
}
\arguments{
\item{a}{events in group 1}

\item{b}{non-events in group 1}

\item{c}{events in group 2}

\item{d}{non-events in group 2}
}
\description{
Calculate log odds ratio
}
