% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{paste2}
\alias{paste2}
\title{Paste 2 values into mean (SD) format}
\usage{
paste2(x, y, d = 2)
}
\arguments{
\item{x}{mean}

\item{y}{SD}

\item{d}{number of digits to display}
}
\value{

}
\description{
Paste 2 values into mean (SD) format
}
