% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{paste3}
\alias{paste3}
\title{Paste 3 values into mean (LCI, UCI) format}
\usage{
paste3(x, y, z, dig = 2)
}
\arguments{
\item{x}{mean}

\item{y}{LCI}

\item{z}{UCI}

\item{dig}{number of digits to display}
}
\value{

}
\description{
Paste 3 values into mean (LCI, UCI) format
}
