% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_par}
\alias{remove_par}
\title{Remove line breaks from character strings}
\usage{
remove_par(x)
}
\arguments{
\item{x}{string}
}
\value{

}
\description{
Remove line breaks from character strings
}
