% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rr}
\alias{rr}
\title{Calculate log risk ratio}
\usage{
rr(a, b, c, d)
}
\arguments{
\item{a}{events in group 1}

\item{b}{non-events in group 1}

\item{c}{events in group 2}

\item{d}{non-events in group 2}
}
\description{
Calculate log risk ratio
}
