% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{st}
\alias{st}
\title{Standardise a vector}
\usage{
st(x, skip_binom = FALSE)
}
\arguments{
\item{x}{vector to standardise (x - mean)/sd}

\item{skip_binom}{if there are only 2 unique values, should they be skipped?}
}
\value{
a standardised vector
}
\description{
Standardise a vector
}
